No longer need to disable the timer when saving TOS
In previous implementations of user functions, the timer interrupt
 was temporarily disabled to ensure that the stack pointer (sp)
could correctly preserve the top of the return address stack (TOS)
 without saving the entire return address stack.
This ensured the TOS value was safely pushed to the stack before
re-enabling the timer interrupt.
Although the duration of timer interrupt disablement was short, it still introduced minor fairness issues and increased the likelihood of user function implementation errors (e.g., forgetting to disable the timer interrupt).

In this commit, the design is changed to treat the return address stack top as if it were a return address register (ra) in other architectures.
While it still requires popping the top of the return address stack after saving it to prevent overflow, this new approach allows us to avoid using timer_disable, at the cost of saving 4 additional bytes of context and adding two extra branch instructions.
Signed-off-by: YU-WEI,HSU <blackfire106.1231@gmail.com>