Store return address into user stack in interrupt
Previously, during an interrupt, three bytes were used to
temporarily save the stack top in order to prevent the
operation of storing the return address into the user
stack from being interrupted.

However, it is actually possible to store these three bytes
 directly into the user stack.
As long as the user function call itself does not cause a
stack overflow, this change will not introduce new risks of
 overflow, while also saving the cost of three temporary
registers.

Although accessing memory through a pointer may increase the
 number of access instructions, in future implementations of
 context switching, pointer-based access will already be
necessary to locate the current context’s storage position.

Therefore, the performance impact in that
scenario will be minimal.

Co-authored-by: Chisheng Chen <johnny1001s000602@gmail.com>
Signed-off-by: YU-WEI,HSU <blackfire106.1231@gmail.com>