Fix method for clearing wait_task_info bits
The previous method for clearing wait_task_info did not handle bit
operations correctly.
It only masked the target bits without shifting the value back to
restore the PID. As a result, any task with a non-zero PID using
the interrupt wait queue would fail.
This commit redesigns the method to ensure that the PID is
correctly restored.

Signed-off-by: YU-WEI,HSU <blackfire106.1231@gmail.com>