Replace get/put char macros with functions
Previously, macros were used to avoid function calls. However,
 according to the XC8 guide, when a function has only a single
char parameter or return value, it is passed via WREG.
Therefore, in this case, using a function call is safe and
correct as long as a stack is available.

The keyboard get implementation remains a macro for now, as
converting it to a function may introduce unknown issues.

Signed-off-by: YU-WEI,HSU <blackfire106.1231@gmail.com>